requirements.txt
apache-airflow-providers-github==2.7.0
apache-airflow-providers-postgres>=5.14.0
sqlalchemy
requests>=2.31.0
pendulum>=2.1.2
pytz
psycopg2-binary>=2.9.9